+++ chaos_theory/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add `make::{permutation, subsequence, subsequence_n, chunks, chunks_n}` and `Source::shuffle`.
+
 ## 0.5.2 (2026-09-07)
 
 - Documentation improvements.